+++ src/HeadingCommand.test.ts
@@ -60,7 +60,8 @@ describe('HeadingCommand', () => {
 
   it('execute without dispatch', () => {
     const state = EditorState.create({ schema: schema1 });
-    command.execute(state);
+    const test = command.execute(state);
+    expect(test).toBeTruthy();
   });
 
   it('execute function() should be return false', () => {
