+++ src/HeadingCommand.ts
@@ -1,7 +1,6 @@
 import { EditorState } from 'prosemirror-state';
 import { Transform } from 'prosemirror-transform';
 import { EditorView } from 'prosemirror-view';
-import noop from './noop';
 import toggleHeading from './toggleHeading';
 import { UICommand } from '@modusoperandi/licit-doc-attrs-step';
 
@@ -22,7 +21,6 @@ class HeadingCommand extends UICommand {
     dispatch?: (tr: Transform) => void,
     _view?: EditorView
   ): boolean => {
-    dispatch = dispatch || noop;
     const { schema, selection } = state;
     const tr = toggleHeading(
       state.tr.setSelection(selection),
@@ -30,7 +28,7 @@ class HeadingCommand extends UICommand {
       this._level
     );
     if (tr.docChanged) {
-      dispatch(tr);
+      dispatch?.(tr);
       return true;
     } else {
       return false;
