+++ src/IndentCommand.test.ts
@@ -49,28 +49,3 @@ describe('IndentCommand', () => {
     expect(command.executeCustom(state, tr, 0, 1)).toBeDefined();
   });
 });
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
