+++ src/TextLineSpacingCommand.ts
@@ -13,7 +13,6 @@ import {
   LINE_SPACING_115,
   LINE_SPACING_150,
 } from './ui/toCSSLineSpacing';
-import noop from './noop';
 
 export function setTextLineSpacing(
   tr: Transform,
@@ -133,7 +132,6 @@ class TextLineSpacingCommand extends UICommand {
     dispatch?: (tr: Transform) => void,
     _view?: EditorView
   ): boolean => {
-    dispatch = dispatch || noop;
     const { schema, selection } = state;
     const tr = setTextLineSpacing(
       state.tr.setSelection(selection),
@@ -141,7 +139,7 @@ class TextLineSpacingCommand extends UICommand {
       this._lineSpacing
     );
     if (tr.docChanged) {
-      dispatch(tr as Transaction);
+      dispatch?.(tr as Transaction);
       return true;
     } else {
       return false;
